RBi - R interface for libbi

rbi is an interface to libbi, a library for Bayesian Inference

Details

The package includes a wrapper for the libbi script, allowing to launch the libbi command from within R. It also provides various utility functions to browse the output from libbi, for instance to plot the results.

The package is made of various components:

A wrapper around libbi called libbi.

A bi_model class that can be used to load and manipulate libbi models

Functions to manipulate the results of the libbi command, which are stored in NetCDF files. Those functions allow to extract variables of interest.

References

http://libbi.org/

See also

Author

Pierre E. Jacob pierre.jacob.work@gmail.com, Anthony Lee awllee@gmail.com, Lawrence Murray lawrence.murray@csiro.au, Sebastian Funk sebastian.funk@lshtm.ac.uk

Examples

## examples for running libbi from rbi (will take a few minutes)
demo(PZ_generate_dataset)
#> 
#> 
#> 	demo(PZ_generate_dataset)
#> 	---- ~~~~~~~~~~~~~~~~~~~
#> 
#> > ### This demo shows how to generate data from a model using bi_generate_dataset
#> > 
#> > # the PZ model file is included in rbi and can be found there:
#> > model_file_name <- system.file(package="rbi", "PZ.bi")
#> 
#> > # assign model variable
#> > PZ <- bi_model(model_file_name)
#> 
#> > # look at the model
#> > PZ
#> bi_model:
#> =========
#>  1: model PZ {
#>  2:   const c = 0.25
#>  3:   const e = 0.3
#>  4:   const m_l = 0.1
#>  5:   const m_q = 0.1
#>  6:   param mu
#>  7:   param sigma
#>  8:   state P
#>  9:   state Z
#> 10:   noise alpha
#> 11:   obs P_obs
#> 12:   sub parameter {
#> 13:     mu ~ uniform(0.0, 1.0)
#> 14:     sigma ~ uniform(0.0, 0.5)
#> 15:   }
#> 16:   sub proposal_parameter {
#> 17:     mu ~ gaussian(mu, 0.02)
#> 18:     sigma ~ gaussian(sigma, 0.01)
#> 19:   }
#> 20:   sub initial {
#> 21:     P ~ log_normal(log(2.0), 0.2)
#> 22:     Z ~ log_normal(log(2.0), 0.1)
#> 23:   }
#> 24:   sub transition(delta = 1.0) {
#> 25:     alpha ~ gaussian(mu, sigma)
#> 26:     ode {
#> 27:       dP/dt = alpha*P - c*P*Z
#> 28:       dZ/dt = e*c*P*Z - m_l*Z - m_q*Z*Z
#> 29:     }
#> 30:   }
#> 31:   sub observation {
#> 32:     P_obs ~ log_normal(log(P), 0.2)
#> 33:   }
#> 34: }
#> 
#> > T <- 100
#> 
#> > # First let's generate a dataset without specifying parameters
#> > # so the parameters and initial conditions are drawn from the prior distribution
#> > # which is specified in the model file.
#> > # if libbi throws an error here, try passing the "working_folder" argument to
#> > # 'bi_generate_dataset' and look for the issue in the ".PZ" subdirectory
#> > dataset1 <- bi_generate_dataset(model=PZ, end_time=T, noutputs=T)
#> Error in locate_libbi(x$path_to_libbi): Could not find libbi executable at '/libbi'
demo(PZ_PMMH)
#> 
#> 
#> 	demo(PZ_PMMH)
#> 	---- ~~~~~~~
#> 
#> > ### This demo shows how to perform filtering on a simple 
#> > ### synthetic dataset using libbi.
#> > # the PZ model file is included in rbi and can be found there:
#> > model_file_name <- system.file(package="rbi", "PZ.bi")
#> 
#> > # assign model variable
#> > PZ <- bi_model(model_file_name)
#> 
#> > # look at the model
#> > PZ
#> bi_model:
#> =========
#>  1: model PZ {
#>  2:   const c = 0.25
#>  3:   const e = 0.3
#>  4:   const m_l = 0.1
#>  5:   const m_q = 0.1
#>  6:   param mu
#>  7:   param sigma
#>  8:   state P
#>  9:   state Z
#> 10:   noise alpha
#> 11:   obs P_obs
#> 12:   sub parameter {
#> 13:     mu ~ uniform(0.0, 1.0)
#> 14:     sigma ~ uniform(0.0, 0.5)
#> 15:   }
#> 16:   sub proposal_parameter {
#> 17:     mu ~ gaussian(mu, 0.02)
#> 18:     sigma ~ gaussian(sigma, 0.01)
#> 19:   }
#> 20:   sub initial {
#> 21:     P ~ log_normal(log(2.0), 0.2)
#> 22:     Z ~ log_normal(log(2.0), 0.1)
#> 23:   }
#> 24:   sub transition(delta = 1.0) {
#> 25:     alpha ~ gaussian(mu, sigma)
#> 26:     ode {
#> 27:       dP/dt = alpha*P - c*P*Z
#> 28:       dZ/dt = e*c*P*Z - m_l*Z - m_q*Z*Z
#> 29:     }
#> 30:   }
#> 31:   sub observation {
#> 32:     P_obs ~ log_normal(log(P), 0.2)
#> 33:   }
#> 34: }
#> 
#> > T <- 50
#> 
#> > init_parameters <- list(P = 2, Z = 2, mu = 0.5, sigma = 0.3)
#> 
#> > # First let's generate a dataset from the model
#> > synthetic_dataset <- bi_generate_dataset(model=PZ, end_time=T, noutputs=T, 
#> +                                          init=init_parameters)
#> Error in locate_libbi(x$path_to_libbi): Could not find libbi executable at '/libbi'
demo(PZ_SMC2)
#> 
#> 
#> 	demo(PZ_SMC2)
#> 	---- ~~~~~~~
#> 
#> > # the PZ model file is included in rbi and can be found there:
#> > model_file_name <- system.file(package="rbi", "PZ.bi")
#> 
#> > # assign model variable
#> > PZ <- bi_model(model_file_name)
#> 
#> > # look at the model
#> > PZ
#> bi_model:
#> =========
#>  1: model PZ {
#>  2:   const c = 0.25
#>  3:   const e = 0.3
#>  4:   const m_l = 0.1
#>  5:   const m_q = 0.1
#>  6:   param mu
#>  7:   param sigma
#>  8:   state P
#>  9:   state Z
#> 10:   noise alpha
#> 11:   obs P_obs
#> 12:   sub parameter {
#> 13:     mu ~ uniform(0.0, 1.0)
#> 14:     sigma ~ uniform(0.0, 0.5)
#> 15:   }
#> 16:   sub proposal_parameter {
#> 17:     mu ~ gaussian(mu, 0.02)
#> 18:     sigma ~ gaussian(sigma, 0.01)
#> 19:   }
#> 20:   sub initial {
#> 21:     P ~ log_normal(log(2.0), 0.2)
#> 22:     Z ~ log_normal(log(2.0), 0.1)
#> 23:   }
#> 24:   sub transition(delta = 1.0) {
#> 25:     alpha ~ gaussian(mu, sigma)
#> 26:     ode {
#> 27:       dP/dt = alpha*P - c*P*Z
#> 28:       dZ/dt = e*c*P*Z - m_l*Z - m_q*Z*Z
#> 29:     }
#> 30:   }
#> 31:   sub observation {
#> 32:     P_obs ~ log_normal(log(P), 0.2)
#> 33:   }
#> 34: }
#> 
#> > T <- 50
#> 
#> > init_parameters <- list(P = 2, Z = 2, mu = 0.5, sigma = 0.3)
#> 
#> > # First let's generate a dataset from the model
#> > synthetic_dataset <- bi_generate_dataset(model=PZ, end_time=T, noutputs=T,
#> +                                          init=init_parameters)
#> Error in locate_libbi(x$path_to_libbi): Could not find libbi executable at '/libbi'
demo(PZ_filtering)
#> 
#> 
#> 	demo(PZ_filtering)
#> 	---- ~~~~~~~~~~~~
#> 
#> > ### This demo shows how to perform filtering on a simple 
#> > ### synthetic dataset using libbi.
#> > # the PZ model file is included in rbi and can be found there:
#> > model_file_name <- system.file(package="rbi", "PZ.bi")
#> 
#> > # assign model variable
#> > PZ <- bi_model(model_file_name)
#> 
#> > # look at the model
#> > PZ
#> bi_model:
#> =========
#>  1: model PZ {
#>  2:   const c = 0.25
#>  3:   const e = 0.3
#>  4:   const m_l = 0.1
#>  5:   const m_q = 0.1
#>  6:   param mu
#>  7:   param sigma
#>  8:   state P
#>  9:   state Z
#> 10:   noise alpha
#> 11:   obs P_obs
#> 12:   sub parameter {
#> 13:     mu ~ uniform(0.0, 1.0)
#> 14:     sigma ~ uniform(0.0, 0.5)
#> 15:   }
#> 16:   sub proposal_parameter {
#> 17:     mu ~ gaussian(mu, 0.02)
#> 18:     sigma ~ gaussian(sigma, 0.01)
#> 19:   }
#> 20:   sub initial {
#> 21:     P ~ log_normal(log(2.0), 0.2)
#> 22:     Z ~ log_normal(log(2.0), 0.1)
#> 23:   }
#> 24:   sub transition(delta = 1.0) {
#> 25:     alpha ~ gaussian(mu, sigma)
#> 26:     ode {
#> 27:       dP/dt = alpha*P - c*P*Z
#> 28:       dZ/dt = e*c*P*Z - m_l*Z - m_q*Z*Z
#> 29:     }
#> 30:   }
#> 31:   sub observation {
#> 32:     P_obs ~ log_normal(log(P), 0.2)
#> 33:   }
#> 34: }
#> 
#> > T <- 100
#> 
#> > init_parameters <- list(P = 2, Z = 2, mu = 0.5, sigma = 0.3)
#> 
#> > # First let's generate a dataset from the model
#> > synthetic_dataset <- bi_generate_dataset(model=PZ, end_time=T,
#> +                                          init=init_parameters)
#> Error in locate_libbi(x$path_to_libbi): Could not find libbi executable at '/libbi'